pangea_securathon/url_checker/requirements.txt
Streamlit==1.21.0
validators==0.20.0
python-whois==0.8.0
